moove all text utils to textElement.ts
