+++ src/element/textElement.ts
@@ -1,14 +1,10 @@
+import { getFontString, arrayToMap } from "../utils";
 import {
-  measureText,
-  getFontString,
-  arrayToMap,
-  getMinCharWidth,
-  wrapText,
-} from "../utils";
-import {
+  ChartType,
   ExcalidrawBindableTextELement,
   ExcalidrawElement,
   ExcalidrawTextElement,
+  FontString,
   NonDeletedExcalidrawElement,
 } from "./types";
 import { mutateElement } from "./mutateElement";
@@ -131,4 +127,257 @@ export const handleBindTextResize = (
       }
     }
   });
+};
+
+// https://github.com/grassator/canvas-text-editor/blob/master/lib/FontMetrics.js
+export const measureText = (
+  text: string,
+  font: FontString,
+  maxWidth?: number | null,
+) => {
+  text = text
+    .split("\n")
+    // replace empty lines with single space because leading/trailing empty
+    // lines would be stripped from computation
+    .map((x) => x || " ")
+    .join("\n");
+  const container = document.createElement("div");
+  container.style.position = "absolute";
+  container.style.whiteSpace = "pre";
+  container.style.font = font;
+
+  if (maxWidth) {
+    const lineHeight = getApproxLineHeight(font);
+    container.style.width = `${String(maxWidth)}px`;
+    container.style.maxWidth = `${String(maxWidth)}px`;
+    container.style.overflow = "hidden";
+    container.style.wordBreak = "break-word";
+    container.style.lineHeight = `${String(lineHeight)}px`;
+    container.style.whiteSpace = "normal";
+  }
+  document.body.appendChild(container);
+  container.innerText = text;
+
+  const span = document.createElement("span");
+  span.style.display = "inline-block";
+  span.style.overflow = "hidden";
+  span.style.width = "1px";
+  span.style.height = "1px";
+  container.appendChild(span);
+  // Baseline is important for positioning text on canvas
+  const baseline = span.offsetTop + span.offsetHeight;
+  const width = container.offsetWidth;
+
+  const height = container.offsetHeight;
+  document.body.removeChild(container);
+
+  return { width, height, baseline };
+};
+
+const DUMMY_TEXT = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789".toLocaleUpperCase();
+export const getApproxLineHeight = (font: FontString) => {
+  return measureText(DUMMY_TEXT, font, null).height;
+};
+
+let canvas: HTMLCanvasElement | undefined;
+const getTextWidth = (text: string, font: FontString) => {
+  if (!canvas) {
+    canvas = document.createElement("canvas");
+  }
+  const canvas2dContext = canvas.getContext("2d") as CanvasRenderingContext2D;
+  canvas2dContext.font = font;
+
+  const metrics = canvas2dContext.measureText(text);
+
+  return metrics.width;
+};
+
+let totalTime = 0;
+let count = 0;
+export const wrapText = (
+  text: string,
+  font: FontString,
+  textContainer: ExcalidrawElement | null,
+) => {
+  if (!textContainer) {
+    return text;
+  }
+  const startTime = performance.now();
+  const maxWidth = textContainer.width - PADDING * 2;
+  console.info(
+    "maxWidth",
+    maxWidth,
+    text.length,
+    "container width",
+    textContainer.width,
+  );
+
+  const lines: Array<string> = [];
+  const originalLines = text.split("\n");
+  originalLines.forEach((originalLine, index) => {
+    const words = originalLine.split(" ");
+    // This means its newline so push it
+    if (words.length === 1 && words[0] === "") {
+      lines.push(words[0]);
+    } else {
+      let currentLine = "";
+      let widthTillNow = 0;
+
+      let index = 0;
+      while (index < words.length) {
+        count++;
+        const currentWidth = getTextWidth(words[index], font);
+
+        if (currentWidth > maxWidth) {
+          widthTillNow = 0;
+          while (words[index].length > 0) {
+            count++;
+
+            const currentChar = words[index][0] as ChartType;
+            const width = charWidth.calculate(currentChar, font);
+            widthTillNow += width;
+            words[index] = words[index].slice(1);
+
+            if (widthTillNow >= maxWidth) {
+              lines.push(currentLine);
+              currentLine = currentChar;
+              widthTillNow = width;
+              if (widthTillNow === maxWidth) {
+                currentLine = "";
+                widthTillNow = 0;
+              }
+            } else {
+              currentLine += currentChar;
+            }
+          }
+          //console.log("words[index]=", words[index]);
+
+          index++;
+          // console.log("index", index);
+        } else {
+          // console.log(
+          //   "width till now",
+          //   widthTillNow,
+          //   "index",
+          //   index,
+          //   "lenght",
+          //   words.length,
+          // );
+          while (widthTillNow < maxWidth && index < words.length) {
+            const word = words[index];
+            widthTillNow = getTextWidth(currentLine + word, font);
+            count++;
+
+            if (widthTillNow >= maxWidth) {
+              lines.push(currentLine);
+              widthTillNow = 0;
+              currentLine = "";
+
+              break;
+            }
+            index++;
+
+            currentLine += `${word} `;
+          }
+          // console.log("index", index, currentLine, words.length);
+
+          if (widthTillNow === maxWidth) {
+            currentLine = "";
+            widthTillNow = 0;
+          }
+        }
+      }
+
+      if (currentLine) {
+        lines.push(currentLine.trim());
+      }
+    }
+  });
+  const endTime = performance.now();
+  const timeTaken = (endTime - startTime) / 1000;
+  totalTime += timeTaken;
+  console.info("Time taken", timeTaken);
+  console.info("total Time taken,", totalTime);
+  console.info("Total runs = ", count);
+  console.info("cacheed char", charWidth.getCache(font));
+  return lines.join("\n");
+};
+
+export const charWidth = (() => {
+  const cachedCharWidth: { [key: FontString]: Array<number> } = {};
+
+  const calculate = (char: ChartType, font: FontString) => {
+    const ascii = char.charCodeAt(0);
+    if (!cachedCharWidth[font]) {
+      cachedCharWidth[font] = [];
+    }
+    if (!cachedCharWidth[font][ascii]) {
+      const width = getTextWidth(char, font);
+      cachedCharWidth[font][ascii] = width;
+    }
+    return cachedCharWidth[font][ascii];
+  };
+
+  const updateCache = (char: ChartType, font: FontString) => {
+    const ascii = char.charCodeAt(0);
+
+    if (!cachedCharWidth[font][ascii]) {
+      cachedCharWidth[font][ascii] = calculate(char, font);
+    }
+  };
+
+  const clearCacheforFont = (font: FontString) => {
+    cachedCharWidth[font] = [];
+  };
+
+  const getCache = (font: FontString) => {
+    return cachedCharWidth[font];
+  };
+  return {
+    calculate,
+    updateCache,
+    clearCacheforFont,
+    getCache,
+  };
+})();
+export const getApproxMinLineWidth = (font: FontString) => {
+  return measureText(DUMMY_TEXT.split("").join("\n"), font).width + PADDING * 2;
+};
+
+export const getApproxMinLineHeight = (font: FontString) => {
+  return getApproxLineHeight(font) + PADDING * 2;
+};
+
+export const getMinCharWidth = (font: FontString) => {
+  const cache = charWidth.getCache(font);
+  if (!cache) {
+    return 0;
+  }
+  const cacheWithOutEmpty = cache.filter((val) => val !== undefined);
+
+  return Math.min(...cacheWithOutEmpty);
+};
+
+export const getApproxCharsToFitInWidth = (font: FontString, width: number) => {
+  // Generally lower case is used so converting to lower case
+  const dummyText = DUMMY_TEXT.toLocaleLowerCase();
+  const batchLength = 6;
+  let index = 0;
+  let widthTillNow = 0;
+  let str = "";
+  while (widthTillNow <= width) {
+    const batch = dummyText.substr(index, index + batchLength);
+    str += batch;
+    widthTillNow += getTextWidth(str, font);
+    if (index === dummyText.length - 1) {
+      index = 0;
+    }
+    index = index + batchLength;
+  }
+
+  while (widthTillNow > width) {
+    str = str.substr(0, str.length - 1);
+    widthTillNow = getTextWidth(str, font);
+  }
+  return str.length;
 };
